Fix .data section generation
So basically everything was backwards. Now the .data section is properly
generated. Variables declared lower and lower in memory. The variables
at the beginning of the line and lower in memory than the ones at the
end. The labels at the addresses of the first variable (lowest) and you
index going up because intrinstically memory is /read/ up by the load
instructions.