Add mul instr, la instr resolving, bugfix with magic check
Add mul instruction to both cpu and instruction. Change la to directly
use datatab for resolving given label instead of resolve() because it's
largely meant to resolve .data section symbols, not .text symbols. So as
of now, it only supports .data section, but in the future it would be
good to make it support both .data and .text labels. This would probably
require unifying the labeltab and datatab into a symtab with both types
of entries. Fix a bug when we check the magic of the file and forgot to
seek back to the beginning when handing off the file handle.