Parse interpreter path correct way using kernel_read
Previously, we just used bprm->buf + hdr->interp_off as the interpreter
string, but that's really the wrong way to do it because we shouldn't be
working directly with the bprm->buf, that's just for header magic
validation and our interpreter just happened to fit into that buffer.
The better way is kernel_read which avoids potential problems if the
interp_off header is malformed, which would cause us to strcpy random
kernel memory later on.