cache: Fix cache read miss block align error
Mask off the offset from the address when reading the block from memory,
so the cache data stays block aligned. Previously, a block was read
starting directly from the address, which resulted in misaligned cache
data.

More on this:

When a cache read miss occurs, that block is fetched from main memory
and placed in the cache. When implementing this the first time, I didn't
quite understand that blocks are aligned, by definition (as defined by
the cache design constants [number of bits for index, offset]). So, when
a miss occurred, I would read a block from memory, starting directly at
the address requested, and place that in the cache. However that address
was likely not block aligned, so a misaligned block would instead be
placed in the cache. This would cause problems further down the road
when future cache reads are attempted because those offset values are
expecting to be relative to an aligned block.